fixed some lingering encode_to_vec calls
